Fix T98735: GPU subdiv displays normals for all elements
The normals flags were not setup properly which made normals for all
elements (vertices, faces) to be drawn when using the normals overlay.
Also remove usage of uints for the flag in the APIs.

+++ source/blender/draw/intern/shaders/common_subdiv_patch_evaluation_comp.glsl
@@ -459,9 +459,9 @@ void main()
     vec3 nor = vec3(0.0);
 
     int origindex = input_vert_origindex[loop_index];
-    uint flag = 0;
+    float flag = 0.0;
     if (origindex == -1) {
-      flag = -1;
+      flag = -1.0;
     }
 
     PosNorLoop vertex_data;
